feat: Add all modules, stored procedures, migrations, and test coverage
- Add IssueMaster, ProductionToStore, LabourChargeInvoice transaction SPs
- Add DeliveryChallan, TaxInvoice SPs and legacy lock system changes
- Add Auth SPs (refresh token, BCrypt, activity log)
- Add all Legacy SPs and migration scripts
- Add C# handlers/commands for all new modules
- Add enterprise lock system (admin controller, cleanup service)
- Add company/financial year repositories with full test coverage
- Update .gitignore to exclude .DS_Store and .claude files

+++ Database_Scripts/StoredProcedures/TaxInvoice/ERP_UpdateTaxInvoice.sql
@@ -4,7 +4,7 @@
 -- Description: Updates an existing Tax Invoice (all fields)
 -- =============================================
 ALTER PROCEDURE [dbo].[ERP_UpdateTaxInvoice]
-    @InvoiceCode INT,
+    @InvoiceCode BIGINT,
     -- (Same parameters as Create, except @InvoiceNumber which doesn't change)
     @CompanyCode INT,
     @InvoiceDate DATETIME,
