+++ ErpBE.Application/TaxInvoice/Commands/DeleteTaxInvoiceCommand.cs
@@ -4,7 +4,7 @@ namespace ErpBE.Application.TaxInvoice.Commands
 {
     public class DeleteTaxInvoiceCommand : IRequest<bool>
     {
-        public int InvoiceCode { get; set; }
+        public long InvoiceCode { get; set; }
         public int CompanyCode { get; set; }
     }
 }
