+++ ErpBE.Application/TaxInvoice/Queries/GetTaxInvoiceByIdQuery.cs
@@ -5,7 +5,7 @@ namespace ErpBE.Application.TaxInvoice.Queries
 {
     public class GetTaxInvoiceByIdQuery : IRequest<TaxInvoiceMasterDto?>
     {
-        public int InvoiceCode { get; set; }
+        public long InvoiceCode { get; set; }
         public int CompanyCode { get; set; }
     }
 }
